backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.6
python-dotenv==1.0.1
alembic==1.13.1
email-validator==2.2.0
Jinja2==3.0.3
